requirements.txt
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
fastapi==0.67.0
greenlet==1.1.0
h11==0.12.0
pycodestyle==2.7.0
pydantic==1.8.2
PyMySQL==1.0.2
PyYAML==5.4.1
pyyml==0.0.2
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
